requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==4.2.17
gunicorn==20.1.0
psycopg2==2.9.10
sqlparse==0.5.2
django-froala-editor